Modified the group functionality to make sure that the user can modify the values in a group he belongs to. Other groups to which he doesn't belong to are hidden when he logs in.
Updated taskpoints and weeklypoints throughout the appln from double to decimal to accomodate 2 precision values.

+++ SourceCode/WeShare.BusinessLogic/BLTaskAssignment.cs
@@ -62,7 +62,7 @@ namespace WeShare.BusinessLogic
             return objDaTaskAssignment.UpdateTaskStatus(taskId, status);
         }
 
-        public bool UpdateTaskPoints(double taskPoints,string userID, int taskId)
+        public bool UpdateTaskPoints(decimal taskPoints,string userID, int taskId)
         {
             return objDaTaskAssignment.UpdateTaskPoints(taskPoints,userID,taskId);
         }
